+++ bench/irmin-pack/trace_stats.ml
@@ -162,11 +162,15 @@ let term_cb =
   in
   Term.(const summary_to_cb $ summary_file)
 
+let deprecated_info = (Term.info [@alert "-deprecated"])
+let deprecated_exit = (Term.exit [@alert "-deprecated"])
+let deprecated_eval_choice = (Term.eval_choice [@alert "-deprecated"])
+
 let () =
   let man = [] in
   let i =
-    Term.info ~man ~doc:"Processing of stat traces and stat trace summaries."
-      "trace_stats"
+    deprecated_info ~man
+      ~doc:"Processing of stat traces and stat trace summaries." "trace_stats"
   in
 
   let man =
@@ -176,7 +180,7 @@ let () =
       `P "trace_stats.exe summarise run0.repr > run0.json";
     ]
   in
-  let j = Term.info ~man ~doc:"Stat Trace to Summary" "summarise" in
+  let j = deprecated_info ~man ~doc:"Stat Trace to Summary" "summarise" in
 
   let man =
     [
@@ -200,10 +204,10 @@ let () =
       `P "trace_stats.exe pp -f r0,run0.json -f r1,run1.repr";
     ]
   in
-  let k = Term.info ~man ~doc:"Comparative Pretty Printing" "pp" in
+  let k = deprecated_info ~man ~doc:"Comparative Pretty Printing" "pp" in
   let l =
-    Term.info ~man ~doc:"Summary JSON to Continous Benchmarks JSON" "cb"
+    deprecated_info ~man ~doc:"Summary JSON to Continous Benchmarks JSON" "cb"
   in
-  Term.exit
-  @@ Term.eval_choice (term_summarise, i)
+  deprecated_exit
+  @@ deprecated_eval_choice (term_summarise, i)
        [ (term_summarise, j); (term_pp, k); (term_cb, l) ]